# compact.Tree: Allow nil visitor and add benchmarks