--- a/integration/log.go
+++ b/integration/log.go
@@ -524,7 +524,7 @@ func buildMemoryMerkleTree(leafMap map[int64]*trillian.LogLeaf, params TestParam
 
 	// We use the leafMap as we need to use the same order for the memory tree to get the same hash.
 	for l := params.StartLeaf; l < params.LeafCount; l++ {
-		if _, err := compactTree.AddLeaf(leafMap[l].LeafValue, func(compact.NodeID, []byte) {}); err != nil {
+		if _, err := compactTree.AddLeaf(leafMap[l].LeafValue, nil); err != nil {
 			return nil, err
 		}
 		if _, _, err := merkleTree.AddLeaf(leafMap[l].LeafValue); err != nil {