This change allows passing nil `visit` function into `AddLeaf` methods,
which makes the code more readable in various places.

Benchmarks don't show significant differences in performance.
Before:
`BenchmarkAddLeafHash-12    	     200	   9425994 ns/op` (varies between 8.8 - 9.6 ms)
After:
`BenchmarkAddLeafHash-12    	     200	   9065164 ns/op` (varies between 8.8 - 9.7 ms)
`BenchmarkAddLeafHashNoVisitor-12    	     200	   9155518 ns/op` (varies between 8.8 - 9.2 ms, slightly faster than the non-nil version)

The benchmarks will also be handy in measuring impact of other
`compact.Tree` optimizations.